refactor: create new controller folder
